+++ examples/basic_compact.js
@@ -35,4 +35,9 @@
   var djiGeozones = new DjiGeozones(opt_options);
   map.addControl(djiGeozones);
 
+  djiGeozones.on('error', (err) => {
+    alert('An error ocurred: ' + err.message);
+    console.log(err)
+  });
+  
 })();